+++ parser/src/main/java/it/auties/protobuf/parser/tree/ProtobufGroupFieldStatement.java
@@ -101,7 +101,7 @@ public final class ProtobufGroupFieldStatement
     }
 
     @Override
-    public Optional<? extends WithIndex> getDirectChildByIndex(int index) {
+    public Optional<? extends WithIndex> getDirectChildByIndex(long index) {
         return ProtobufStatementWithBodyImpl.getDirectChildByIndex(children, index);
     }
 
@@ -111,7 +111,7 @@ public final class ProtobufGroupFieldStatement
     }
 
     @Override
-    public <V extends ProtobufTree> Optional<? extends V> getDirectChildByIndexAndType(int index, Class<V> clazz) {
+    public <V extends ProtobufTree> Optional<? extends V> getDirectChildByIndexAndType(long index, Class<V> clazz) {
         return ProtobufStatementWithBodyImpl.getDirectChildByIndexAndType(children, index, clazz);
     }
