+++ parser/src/main/java/it/auties/protobuf/parser/tree/ProtobufOneofFieldStatement.java
@@ -96,7 +96,7 @@ public final class ProtobufOneofFieldStatement
     }
 
     @Override
-    public Optional<? extends WithIndex> getDirectChildByIndex(int index) {
+    public Optional<? extends WithIndex> getDirectChildByIndex(long index) {
         return ProtobufStatementWithBodyImpl.getDirectChildByIndex(children, index);
     }
 
@@ -106,7 +106,7 @@ public final class ProtobufOneofFieldStatement
     }
 
     @Override
-    public <V extends ProtobufTree> Optional<? extends V> getDirectChildByIndexAndType(int index, Class<V> clazz) {
+    public <V extends ProtobufTree> Optional<? extends V> getDirectChildByIndexAndType(long index, Class<V> clazz) {
         return ProtobufStatementWithBodyImpl.getDirectChildByIndexAndType(children, index, clazz);
     }
